requirements.txt
streamlit
wikipedia
groq
requests
Pillow
regex
python-dotenv
edge-tts
moviepy
imageio
imageio-ffmpeg
numpy
pydub
soundfile
ffmpeg-python
google-genai